Add files via upload
The purpose of this code is to predict the initial values of q, deltaD, deltaO18, and dexcess inside the bag. These initial values correspond to the true values or, in the case of profiles, the corrected values.

The code numerically solves the differential equations for q(t) and R(t), but the time is reversed compared to the evolution_bag.pl process.
The results for q0, deltaO180, and d0 are written as the last three numbers in the final line of the output file.